In the middle of a few things.  Currently working on:
1)  Converting Markers to Custer Items
2)  Moving from POST to PUT

PUT is working on the initial commit of dot.  I will need to work on getting the proper IDs from JK

The Cluster Items are going to be a little more involved.  It will require converting markers to CIs, changing the icon for beforeCluster to the old drawables (maybe making new drawables all together), and editing at what zoom things begin to cluster.